web: Do not truncate overall rating to an integer
Now the rating is left as a floating point number so that the
templates can choose the number of decimal places to display.

+++ lib/hedwig/type.py
@@ -903,7 +903,7 @@ class ReviewerCollection(ResultCollection):
         if not total_weight:
             return None
 
-        return int(total_rating / total_weight)
+        return total_rating / total_weight
 
     def get_person(self, person_id, roles=None):
         for member in self.values():
